put describePagination into its own file
